moved Prop 2

+++ Common/src/Bytes.cpp
@@ -110,7 +110,7 @@ uint32_t Bytes::length() const {
 int Bytes::offset(int32_t pos) {
 	if (pos < 0)
 		_offset = _limit;
-	else if (pos < _capacity)
+	else if ((uint32_t)pos < _capacity)
 		_offset = pos;
 	return 0;
 }
